# docs/swagger.yaml
openapi: 3.0.0
info:
  description: |
    This service is responsible for
    
    - Routing requests from front to backend services (smth like BFF pattern)
    - Enriching responses
    - Preparing data for the front (change types etc.)
    - Request validation from front (so back will not be loaded).
  version: "1.0.0"
  title: API-Gateway
paths:
  '/v1/users/patients/{doctor_id}':
    get:
      tags:
        - patients
      summary: Get all patients of the doctor
      description: | 
        Returns ids of all the patients, who have treatment(s) conducting by the given doctor.
        
        If there is more than one traetment for the one given patient- id is returned only ones.
        
        Method is being proxied to process_execution_service
      parameters:
        - name: doctor_id
          in: path
          description: ID of the doctor
          required: true
          schema:
            type: string
      operationId: getAllPatients
      responses:
        '404':
          description: No doctor found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                    type: string
  # '/v1/users/patients/{patient_id}':
  #   get:
  #     tags:
  #       - patients
  #     summary: Get info about patients by ids
  #     parameters:
  #       - name: patient_id
  #         in: path
  #         description: ID of the patient
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '404':
  #         description: No user found
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                   type: object
  '/v1/instances/{patient_id}':
    get:
      tags:
        - instances
      summary: Get all running treatment instances of the patient
      description: |
        Method return short info about all the treatments of the given patient
        
        Different patients cannot have the same treatment ID
      
        Method is being proxied to process_execution_service
      parameters: 
        - name: patient_id
          in: path
          description: ID of the patient 
          required: true
          schema:
            type: string
      operationId: getPatientTreatments
      responses:
        '404':
          description: No patient found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties: 
                  patient_id:
                    type: string
                  treatments:
                    type: array
                    items:
                      $ref: '#/components/schemas/TreatmentLight'
  '/v1/instance/id={id}':
    get:
      tags:
        - instances
      summary: Get running instance by ID
      description: |
        Method return the full treatment object
        
        Method is being proxied to process_execution_service
      parameters:
        - name: id
          in: path
          description: ID of the instance
          required: true
          schema:
            type: string
      responses:
        '404':
          description: No instance found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                    $ref: '#/components/schemas/Treatment'
  # '/v1/instances/status/id={id}':
  #   get:
  #     tags:
  #       - instances
  #     summary: Get status of the running instance by ID
  #     parameters:
  #       - name: id
  #         in: path
  #         description: ID of the instance
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '404':
  #         description: No user found
  #       '405':
  #         description: No instance found
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             schema:
  #                 $ref: '#/components/schemas/TreatmentStatus'
  '/v1/instances/create':
    post:
      tags:
        - instances
      summary: Create treatment from the schema
      description: |
        Method gets schema from schema_service
        
        Then goes to process_execution_service
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                schema_id:
                  type: string
                patient_id:
                  type: string
                doctor_id:
                  type: string
      responses:
        '404':
          description: No schema found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Treatment'
  # '/v1/instance/tasks':
  #   get:
  #     tags:
  #       - tasks
  #     summary: Get children of some Instance/Task
  #     parameters:
  #       - name: instance_id
  #         in: query
  #         description: ID of the instance
  #         required: true
  #         schema:
  #           type: string
  #       - name: task_id
  #         in: query
  #         description: ID of the task
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '404':
  #         description: No user found
  #       '405':
  #         description: No instance/task found
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             schema:
  #                 $ref: '#/components/schemas/Task'
  # '/v1/instance/tasks/is_locked':
  #   get:
  #     tags:
  #       - tasks
  #     summary: If the task is locked
  #     parameters:
  #       - name: instance_id
  #         in: query
  #         description: ID of the instance
  #         required: true
  #         schema:
  #           type: string
  #       - name: task_id
  #         in: query
  #         description: ID of the task
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '404':
  #         description: No user found
  #       '405':
  #         description: No instance/task found
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             schema:
  #                 type: object
  #                 properties:
  #                   is_locked:
  #                     type: boolean
  # '/v1/instance/tasks/status':
  #   get:
  #     tags:
  #       - tasks
  #     summary: Get status of the Task
  #     parameters:
  #       - name: instance_id
  #         in: query
  #         description: ID of the instance
  #         required: true
  #         schema:
  #           type: string
  #       - name: task_id
  #         in: query
  #         description: ID of the task
  #         required: true
  #         schema:
  #           type: string
  #     responses:
  #       '404':
  #         description: No user found
  #       '405':
  #         description: No instance/task found
  #       '200':
  #         description: Successful operation
  #         content:
  #           application/json:
  #             schema:
  #                 $ref: '#/components/schemas/TaskStatus'
  '/v1/instance/tasks/complete':
    post:
      tags:
        - tasks
      summary: Complete the Task in the treatment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                instance_id:
                  type: string
                task_ids:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '404':
          description: No instance found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: integer
                      format: int64
                    task_status:
                      $ref: '#/components/schemas/TaskStatus'
  '/v1/schemas':
    get:
      tags:
        - schemas
      summary: Get all the available schemas
      responses:
        '404':
          description: No user found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
  '/v1/schemas/{schema_id}':
    get:
      tags:
        - schemas
      summary: Get the schema by ID
      description: |
        Return schema from schema_service by ID
      parameters: 
        - name: schema_id
          in: path
          description: ID of the schema 
          required: true
          schema:
            type: string
      responses:
        '404':
          description: No schema found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
  '/v1/schema/update':
    post:
      tags:
        - schemas
      summary: Update a schema
      description: |
        This method udpates a schema by id
        
        Returns full schema object
        
        Method is proxied to schema_service
      requestBody:
        content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
      responses:
        '404':
          description: No user found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
  '/v1/schema/create':
    post:
      tags:
        - schemas
      summary: Create a schema
      description: | 
        This method creates a schema
        
        IDs for each task should be generated at frontend
        
        Returns full schema object
        
        Method is proxied to schema_service
      requestBody:
        content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaLight'
      responses:
        '404':
          description: No user found
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
components:
  schemas:
    TaskStatus:
      type: string
      enum:
        - UNSPECIFIED
        - NOT_STARTED
        - IN_PROGRESS
        - BLOCKED
        - DONE
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        level:
          type: integer
        name:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        blocked_by:
          type: array
          items:
            type: integer
            format: int64
        responsible:
          type: string
        time_limit:
          type: integer
          format: int64
        childrens:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        comment:
          type: string
    SchemaLight:
      type: object
      properties:
        author_id:
          type: string
        doctor_id:
          type: string
        schema_name:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    Schema:
      type: object
      properties:
        schema_id:
          type: string
        author_id:
          type: integer
        schema_name:
          type: string
        created_at:
          type: string 
          format: date-time
        updated_at:
          type: string 
          format: date-time
        deleted_at:
          type: string 
          format: date-time
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    SchemaInstance:
      type: object
      properties:
        instance_id:
          type: string
        schema_reference:
          $ref: '#/components/schemas/Schema'
    PatternInstance:
      type: object
      properties:
        instance_id:
          type: string
        status:
          type: string
          enum: 
            - PATTERN_INSTANCE_STATUS_UNSPECIFIED
            - PATTERN_INSTANCE_STATUS_NOT_STARTED
            - PATTERN_INSTANCE_STATUS_RUNNING
            - PATTERN_INSTANCE_STATUS_BLOCKED
            - PATTERN_INSTANCE_STATUS_COMPLETED
            - PATTERN_INSTANCE_STATUS_CANCELLED
        pattern_id:
          type: string
        author_id:
          type: string
        pattern_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    TreatmentStatus:
      type: string
      enum:
        - UNSPECIFIED
        - NOT_STARTED
        - IN_PROGRESS
        - BLOCKED
        - DONE
    Treatment:
      type: object
      properties:
        treatment_id:
          type: string
        doctor_id:
          type: string
        patient_id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/TreatmentStatus'
        pattern_instance:
          $ref: '#/components/schemas/PatternInstance'
        started_at:
          type: string 
          format: date-time
        finished_at:
          type: string 
          format: date-time
        deleted_at:
          type: string 
          format: date-time
    TreatmentLight:
      type: object
      properties:
        treatment_id:
          type: string
          description: Treatment ID
        treatment_name:
          type: string
          description: Treatment Name
        status:
          $ref: '#/components/schemas/TreatmentStatus'
          description: Treatment Stats
        treatment_progress:
          type: number
          description: Process of the treatment completion
        
        
            
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/GlobalLabsAltTeam/API-Gateway/1.0.0